New: base templates in a single file
Previously the header/footer files were concatenated into the built file. If there were complex layouts with divs that needed to be opened, this meant that closing divs had to go in a separate file.

This commit changes this to have all of the 'base' template in a single file, 'base.html', and the templates are replaced with a formatting variable in 'base.html', '{templates}'.